+++ src/ffi/intrcheck.rs
@@ -4,10 +4,7 @@ extern "C" {
     #[cfg_attr(PyPy, link_name = "PyPyOS_InterruptOccurred")]
     pub fn PyOS_InterruptOccurred() -> c_int;
     #[cfg(not(Py_3_10))]
-    #[deprecated(
-        since = "0.14.0",
-        note = "Not documented in Python API; see Python 3.10 release notes"
-    )]
+    #[deprecated(note = "Not documented in Python API; see Python 3.10 release notes")]
     pub fn PyOS_InitInterrupts();
 
     #[cfg(any(not(Py_LIMITED_API), Py_3_7))]